refactor: change code structure based on FSD
