+++ src/app/layouts/MainLayout.tsx
@@ -1,10 +1,10 @@
 import { type FC } from 'react';
 import { Outlet } from 'react-router-dom';
 
-import { Flyout } from './Flyout.tsx';
-import { Header } from './Header.tsx';
+import { Flyout } from '../../svalka/Flyout.tsx';
+import { Header } from '../../svalka/Header.tsx';
 
-import type { Character } from '../utils/types/rickAndMorty.ts';
+import type { Character } from '@/utils/types/rickAndMorty.ts';
 
 interface MainLayoutProps {
   handleSearch: (query: string) => void;
@@ -13,13 +13,7 @@ interface MainLayoutProps {
 
 export const MainLayout: FC<MainLayoutProps> = ({ handleSearch }) => {
   return (
-    <div
-      className="min-h-screen"
-      style={{
-        backgroundColor: 'var(--color-bg)',
-        color: 'var(--text-primary)',
-      }}
-    >
+    <div className="min-h-screen bg-bg text-text-primary">
       <Header handleSearch={handleSearch} />
       <Outlet />
       <Flyout />
