+++ src/entities/character/CharacterDetails.tsx
