+++ src/features/search-character/SearchBar.tsx
