+++ src/shared/close-button/CloseButton.tsx
